split app.py files by funtions
