+++ templates/menu.html
@@ -33,18 +33,18 @@
         </style>
 </head>
   <body>
-    <form id="logout" action="/logout" method="POST">
+    <form id="logout" action="/auth/logout" method="POST">
     <div style="padding-left: 90%; margin: 30px;">
     <button type="submit" class="button is-link is-outlined">로그아웃</button>
 </div>
 </form>
-<form id="study" action="/study" method="POST">
+<form id="study" action="/studybutton" method="POST">
     <div class="wrap">
         <div class="btn_wrap">
             <button type="submit" class="large">공부 시작!</button>
     </div>
   </form>
-    <form id="rank" action="/rank" method="POST">
+    <form id="rank" action="/rankbutton" method="POST">
     <div class="btn_wrap">
         <button type="submit" class="large">순위 보기</button>
     </div>
